+++ src/components/calendar/calendar.js
@@ -32,6 +32,12 @@ const COMPONENT_NAME_DEFAULTS = {
   modalTemplate: null,
   menuId: null,
   menuSelected: null,
+  showTooltip: true,
+  tooltipSettings: {
+    subject: true,
+    comments: true,
+    time: true,
+  },
   eventTooltip: 'overflow',
   iconTooltip: 'overflow',
   newEventDefaults: {
@@ -95,6 +101,10 @@ const COMPONENT_NAME_DEFAULTS = {
  * @param {string} [settings.menuId=null] ID of the menu to use for an event right click context menu
  * @param {function} [settings.menuSelected=null] Callback for the  right click context menu
  * @param {string} [settings.newEventDefaults] Initial event properties for the new events dialog.
+ * @param {boolean} [settings.showTooltip] If true, displays tooltip on event mouse hover.
+ * @param {boolean} [settings.tooltipSettings.subject] If true, displays event subject on tooltip.
+ * @param {boolean} [settings.tooltipSettings.comments] If true, displays event comments on tooltip.
+ * @param {boolean} [settings.tooltipSettings.time] If true, displays event start and end time on tooltip.
  * @param {string | function} [settings.eventTooltip] The content of event tooltip. Default value is 'overflow'
  * @param {string | function} [settings.iconTooltip] The content of event icon tooltip. Default value is 'overflow'
  * @param {boolean} [settings.showToday=true] Deterimines if the today button should be shown.
@@ -304,7 +314,9 @@ Calendar.prototype = {
       month: this.settings.month,
       year: this.settings.year,
       day: this.settings.day,
-      eventTooltip: this.eventTooltip,
+      showTooltip: this.settings.showTooltip,
+      tooltipSettings: this.settings.tooltipSettings,
+      eventTooltip: this.settings.eventTooltip,
       iconTooltip: this.iconTooltip,
       showToday: this.settings.showToday,
       showViewChanger: this.settings.showViewChanger,
@@ -384,6 +396,8 @@ Calendar.prototype = {
       showViewChanger: this.settings.showViewChanger,
       hitbox: this.settings.hitbox,
       onChangeView: this.onChangeToWeekDay,
+      showTooltip: this.settings.showTooltip,
+      tooltipSettings: this.settings.tooltipSettings,
       eventTooltip: this.settings.eventTooltip,
       iconTooltip: this.settings.iconTooltip,
       attributes: this.settings.attributes,
@@ -393,7 +407,7 @@ Calendar.prototype = {
 
     this.weekView.settings.filteredTypes = this.filterEventTypes();
     this.weekView.settings.onChangeWeek = (args) => {
-      this.monthView.selectDay(args.startDate, false, true);
+      this.monthView.selectDay(args.startDate, false, true, null, args.endDate);
     };
     this.weekView.renderAllEvents();
     utils.addAttributes($('.calendar-weekview'), this, this.settings.attributes, 'weekview');
@@ -947,7 +961,7 @@ Calendar.prototype = {
     }
 
     node.innerHTML = `<div class="calendar-event-content">
-      ${event.icon ? `<span class="calendar-event-icon"><svg class="icon ${event.icon}" focusable="false" aria-hidden="true" role="presentation" data-status="${event.status}"><use href="#${event.icon}"></use></svg></span>` : ''}
+      ${event.icon ? `<span class="calendar-event-icon"><svg class="icon ${event.icon}" focusable="false" aria-hidden="true" role="presentation" data-status="${event.status !== undefined ? event.status : ''}"><use href="#${event.icon}"></use></svg></span>` : ''}
       <span class="calendar-event-title">${event.shortSubject || event.subject}</span>
     </div>`;
 
@@ -990,17 +1004,31 @@ Calendar.prototype = {
       }
     }
 
-    if (this.settings.eventTooltip !== 'overflow') {
-      if (typeof this.settings.eventTooltip === 'function') {
-        this.settings.eventTooltip({
-          month: this.settings.month,
-          year: this.settings.year,
-          event
-        });
-      } else if (event[this.settings.eventTooltip]) {
-        node.setAttribute('title', event[this.settings.eventTooltip]);
+    if (this.settings.showTooltip) {
+      if (this.settings.eventTooltip !== 'overflow') {
+        if (typeof this.settings.eventTooltip === 'function') {
+          this.settings.eventTooltip({
+            month: this.settings.month,
+            year: this.settings.year,
+            event
+          });
+        } else if (event[this.settings.eventTooltip]) {
+          node.setAttribute('title', event[this.settings.eventTooltip]);
+          $(node).tooltip({
+            content: node.innerText
+          });
+        }
+      } else {
+        const s = this.settings;
+        const tooltipContent = `
+          ${s.tooltipSettings.subject ? `${event.subject} <br>` : ''}
+          ${s.tooltipSettings.comments ? `${event.comments} <br>` : ''}
+          ${s.tooltipSettings.time ? `from ${event.startsHourLocale} to ${event.endsHourLocale}` : ''}
+          `;
+        node.setAttribute('title', tooltipContent);
         $(node).tooltip({
-          content: node.innerText
+          content: node.innerText,
+          contentAlignment: 'left'
         });
       }
     }